Please read extended description
So to make the process easy. I was able to take all my forms I generated and transfer them to Jeremy original log-in sln. That will create less headache then trying to transfer everything to a new sln. So eveyone should be able to pull it and edited from the original sln.